Move the setup() test function into a helpers file
